move mmh3 hash for reusability
